# contracts/api.yaml
openapi: 3.1.0
info:
  title: ClipNotes API
  version: 0.2.0
  description: |
    REST contract for clip registration, Hafnia-triggered analysis, monitoring data, and reasoning endpoints
    used by the ClipNotes application.
servers:
  - url: http://localhost:8000
paths:
...
  /api/reasoning/compare:
    post:
      summary: Compare two analyzed clips and return a normalized reasoning answer.
      tags: [reasoning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasoningCompareRequest'
      responses:
        '200':
          description: Comparison response with evidence and metrics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasoningComparisonResponse'
        '400':
          description: Validation error or missing analysis data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reasoning/chat:
    post:
      summary: Ask follow-up questions about one or more clips.
      tags: [reasoning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasoningChatRequest'
      responses:
        '200':
          description: Chat response with reasoning answer and evidence references.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasoningChatResponse'
        '400':
          description: Validation error or missing analysis data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reasoning/history:
    get:
      summary: Retrieve the most recent reasoning exchanges for selected clips.
      tags: [reasoning]
      parameters:
        - in: query
          name: clip_id
          schema:
            type: string
            format: uuid
        - in: query
          name: clip_selection_hash
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: History list ordered by newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasoningHistoryResponse'
...
components:
  schemas:
...
    ReasoningCompareRequest:
      type: object
      required: [clip_a, clip_b, question]
      properties:
        clip_a:
          type: string
          format: uuid
        clip_b:
          type: string
          format: uuid
        question:
          type: string
          minLength: 1
    ReasoningComparisonResponse:
      type: object
      required: [answer, explanation, evidence]
      properties:
        answer:
          type: string
          enum: [clip_a, clip_b, equal, uncertain]
        explanation:
          type: string
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/ReasoningEvidence'
        metrics:
          $ref: '#/components/schemas/ReasoningMetricsCore'
        confidence:
          type: number
          minimum: 0
          maximum: 1
    ReasoningChatRequest:
      type: object
      required: [clips, message]
      properties:
        clips:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        message:
          type: string
          minLength: 1
    ReasoningChatResponse:
      type: object
      required: [answer, created_at]
      properties:
        answer:
          type: string
        created_at:
          type: string
          format: date-time
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/ReasoningEvidence'
        clips:
          type: array
          items:
            type: string
            format: uuid
    ReasoningHistoryResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReasoningHistoryEntry'
    ReasoningHistoryEntry:
      type: object
      required: [id, clip_ids, question, answer, created_at, answer_type]
      properties:
        id:
          type: string
          format: uuid
        clip_ids:
          type: array
          items:
            type: string
            format: uuid
        question:
          type: string
        answer:
          $ref: '#/components/schemas/ReasoningChatResponse'
        answer_type:
          type: string
          enum: [comparison, chat]
        created_at:
          type: string
          format: date-time
...
    ReasoningMetricsCore:
      type: object
      properties:
        counts_by_label:
          type: object
          additionalProperties:
            type: integer
        severity_distribution:
          type: object
          additionalProperties:
            type: number
...
    ReasoningEvidence:
      type: object
      required: [clip_id, label]
      properties:
        clip_id:
          type: string
          format: uuid
        label:
          type: string
        timestamp_range:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
        description:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
  securitySchemes: {}
...